optional-requirements.txt
# GDrive Integration
google-api-python-client>=2.73.00,<2.200.0
gevent>20.6.0,<25.9.2
greenlet>=0.4.17,<3.4.0
httplib2>=0.9.2,<0.32.0
oauth2client>=4.0.0,<4.1.4
uritemplate>=3.0.0,<4.3.0
pyasn1-modules>=0.0.8,<0.7.0
pyasn1>=0.1.9,<0.7.0
PyDrive2>=1.15.0,<1.22.0
PyYAML>=3.12,<6.1
rsa>=3.4.2,<4.10.0

# Gmail
google-auth-oauthlib>=1.0.0,<1.3.0
google-api-python-client>=2.73.00,<2.200.0

# goodreads
goodreads>=0.3.2,<0.4.0
python-Levenshtein>=0.12.0,<0.28.0

# ldap login
python-ldap>=3.0.0,<3.5.0
Flask-SimpleLDAP>=1.4.0,<2.2.0

# oauth
Flask-Dance>=2.0.0,<7.2.0
SQLAlchemy-Utils>=0.33.5,<0.43.0

# metadata extraction
rarfile>=3.2,<5.0
scholarly>=1.2.0,<1.8
markdown2>=2.0.0,<2.6.0
html2text>=2020.1.16,<2025.4.16
python-dateutil>=2.1,<2.10.0
beautifulsoup4>=4.0.1,<4.15.0
faust-cchardet>=2.1.18,<2.1.20
py7zr>=0.15.0,<0.21.0
mutagen>=1.40.0,<1.50.0
pycountry>=20.0.0,<25.0.0

# Comics
natsort>=2.2.0,<8.5.0
comicapi>=2.2.0,<3.3.0

# Kobo integration
jsonschema>=3.2.0,<4.30.0

# S3 Integration
boto3>=1.26.0,<1.37.0
